Move get hooks to getAccountDataHooks
