+++ src/components/AccountDrawer/RenameAccountModal.tsx
@@ -14,7 +14,7 @@ import {
   ModalContent,
   ModalFooter,
 } from "@chakra-ui/react";
-import { useAllAccounts } from "../../utils/hooks/setAccountDataHooks";
+import { useAllAccounts } from "../../utils/hooks/getAccountDataHooks";
 import { useContactExists } from "../../utils/hooks/contactsHooks";
 import FormPageHeader from "../SendFlow/FormPageHeader";
 import renameAccount from "../../utils/redux/thunks/renameAccount";
