+++ src/components/ContactModal.tsx
@@ -13,18 +13,18 @@ import {
   ModalHeader,
   Text,
 } from "@chakra-ui/react";
-import { FC, useEffect, useRef, useContext } from "react";
+import { FC, useContext, useEffect, useRef } from "react";
 import { useForm } from "react-hook-form";
 import colors from "../style/colors";
 import { isAddressValid } from "../types/Address";
 import { Contact } from "../types/Contact";
-import { useAllAccounts, useGetOwnedAccountSafe } from "../utils/hooks/setAccountDataHooks";
 import { useContactExists } from "../utils/hooks/contactsHooks";
+import { useAllAccounts, useGetOwnedAccountSafe } from "../utils/hooks/getAccountDataHooks";
 import { useAppDispatch } from "../utils/redux/hooks";
 import { contactsActions } from "../utils/redux/slices/contactsSlice";
 import { CopyableAddress } from "./CopyableText";
-import { FormErrorMessage } from "./FormErrorMessage";
 import { DynamicModalContext } from "./DynamicModal";
+import { FormErrorMessage } from "./FormErrorMessage";
 
 export const UpsertContactModal: FC<{
   title: string;
