+++ src/components/Onboarding/nameAccount/NameAccount.tsx
@@ -1,5 +1,5 @@
 import { NameAccountStep, Step, StepType } from "../useOnboardingModal";
-import { useIsUniqueLabel } from "../../../utils/hooks/setAccountDataHooks";
+import { useIsUniqueLabel } from "../../../utils/hooks/getAccountDataHooks";
 import NameAccountDisplay from "./NameAccountDisplay";
 
 export const DEFAULT_ACCOUNT_LABEL = "Account";
