+++ src/views/home/AccountsList.tsx
@@ -15,11 +15,7 @@ import { DynamicModalContext } from "../../components/DynamicModal";
 import NestedScroll from "../../components/NestedScroll";
 import { useOnboardingModal } from "../../components/Onboarding/useOnboardingModal";
 import { Account } from "../../types/Account";
-import {
-  useAllAccounts,
-  useRemoveMnemonic,
-  useRemoveNonMnemonic,
-} from "../../utils/hooks/setAccountDataHooks";
+import { useRemoveMnemonic, useRemoveNonMnemonic } from "../../utils/hooks/setAccountDataHooks";
 import { useAsyncActionHandler } from "../../utils/hooks/useAsyncActionHandler";
 import { useAppDispatch, useAppSelector } from "../../utils/redux/hooks";
 import { deriveAccount } from "../../utils/redux/thunks/restoreMnemonicAccounts";
@@ -30,6 +26,7 @@ import { AccountTile } from "../../components/AccountTile/AccountTile";
 import colors from "../../style/colors";
 import { ConfirmationModal } from "../../components/ConfirmationModal";
 import AddAccountIcon from "../../assets/icons/AddAccount";
+import { useAllAccounts } from "../../utils/hooks/getAccountDataHooks";
 
 export const AccountListHeader = () => {
   const { onOpen, modalElement } = useOnboardingModal();
